+++ R/01-fitModel.R
@@ -6,44 +6,16 @@ utils::globalVariables(c("standardizedX", "Date3"))
 #' Get plotValues
 #'
 #' @param plotValues (list) list with all values required for the plot
-#' @param activeFile (character) name of the selected file
-#' @param activeFileData (data.frame) content of the selected file
-#' @param dataSelection (list) list of reactive selected columns of the
-#'  current file
 #' @param modelParameters (list) list of settings for the model
-getPlotValues <- function(plotValues, activeFile, activeFileData, dataSelection,
-                          modelParameters){
+fitModel <- function(plotValues,
+                            modelParameters){
   plotValues$modelData <- NULL
   plotValues$predictedData <- NULL
-
-  plotValues$activeFile <- activeFile
-  plotValues$activeFileData <- activeFileData
-  plotValues$dataSettings <- list(xColumns = dataSelection$xColumns(),
-                                  yColumns = dataSelection$yColumns(),
-                                  dataOutlier = dataSelection$dataOutlier())
 
   xSelection <- getSelection(plotValues$dataSettings$xColumns)
   ySelection <- getSelection(plotValues$dataSettings$yColumns)
 
-  plotValues$selectedData <-
-    activeFileData[, unlist(c(xSelection$colNames, ySelection$colNames))]
-
-  if (any(!sapply(plotValues$selectedData, is.numeric))) {
-    # transform to numeric
-    plotValues$selectedData <- toNumericCols(plotValues$selectedData)
-  }
-
-  if (!is.null(plotValues$selectedData) && any(is.na(plotValues$selectedData))) {
-    plotValues$selectedData <- na.omit(plotValues$selectedData)
-  }
-
   if (is.null(plotValues$selectedData)) return(plotValues)
-
-  plotValues$selectedData <- addColumnDataOutlier(
-    selectedData = plotValues$selectedData,
-    ySelection = ySelection,
-    dataOutlier = plotValues$dataSettings$dataOutlier
-  )
 
   prepData <- getPrepData(data = plotValues$selectedData,
                           xSelection = xSelection,
@@ -62,8 +34,10 @@ getPlotValues <- function(plotValues, activeFile, activeFileData, dataSelection,
 
   plotValues$predictedData <- predictData(modelData = plotValues$modelData$modelOutput,
                                prepData = prepData,
-                               smoothConst = plotValues$modelParameters$smoothConst)
+                               smoothConst = plotValues$modelParameters$smoothConst) %>%
+    tryCatchWithWarningsAndErrors(errorTitle = "Prediction failed", alertStyle = "shinyalert")
 
+  req(!is.null(plotValues$predictedData))
   plotValues$defaultXRange <- getRange(
     data = plotValues$selectedData[, unlist(xSelection$colNames),
                                    drop = FALSE],
@@ -77,17 +51,22 @@ getPlotValues <- function(plotValues, activeFile, activeFileData, dataSelection,
 }
 
 getModelFit <- function(data,
-                     prepData,
-                     xSelection, ySelection,
-                     modelParameters,
-                     isCheck = FALSE
-){
-  modelOutput <- fitModel(
-    prepData = prepData,
-    modelParameters = modelParameters,
-    progressMessage = "Calculating Model",
-    isCheck = isCheck
-  )
+                        prepData,
+                        xSelection, ySelection,
+                        modelParameters){
+  modelOutput <- prepData %>%
+    fitPlotRModelMC(K = modelParameters$K,
+                    burnin = modelParameters$burnin,
+                    iter = modelParameters$iter,
+                    penalty = modelParameters$const,
+                    smoothConst = modelParameters$smoothConst,
+                    nChains = modelParameters$nChains,
+                    sdVar = modelParameters$sdVar,
+                    progressMessage = "Calculating Model"
+    ) %>%
+    tryCatchWithWarningsAndErrors(errorTitle = "Modeling failed",
+                                  warningTitle = "Warning",
+                                  alertStyle = "shinyalert")
 
   if (is.null(modelOutput)) {
     data$isModelOutlier <- FALSE
@@ -99,20 +78,32 @@ getModelFit <- function(data,
 
     predictedData <- predictData(modelData = modelOutput,
                                  prepData = prepData,
-                                 smoothConst = modelParameters$smoothConst)
+                                 smoothConst = modelParameters$smoothConst) %>%
+      tryCatchWithWarningsAndErrors(errorTitle = "Prediction failed", alertStyle = "shinyalert")
 
+    req(!is.null(predictedData))
     data <- findModelOutlier(data = data,
                              predictedData = predictedData$observations,
                              yNames = ySelection$colNames,
                              yType = ySelection$type,
                              outlierValue = modelParameters$outlierValue)
 
-    modelOutput <- fitModel(
-      prepData = getPrepData(data = data,
-                             xSelection = xSelection, ySelection = ySelection),
-      modelParameters = modelParameters,
-      progressMessage = "Removing model outliers",
-      isCheck = isCheck)
+    modelOutput <- data %>%
+      getPrepData(xSelection = xSelection,
+                  ySelection = ySelection) %>%
+      fitPlotRModelMC(
+        K = modelParameters$K,
+        burnin = modelParameters$burnin,
+        iter = modelParameters$iter,
+        penalty = modelParameters$const,
+        smoothConst = modelParameters$smoothConst,
+        nChains = modelParameters$nChains,
+        sdVar = modelParameters$sdVar,
+        progressMessage = "Removing model outliers"
+      ) %>%
+      tryCatchWithWarningsAndErrors(errorTitle = "Modeling failed",
+                                    warningTitle = "Warning",
+                                    alertStyle = "shinyalert")
 
     list(data = data,
          modelOutput = modelOutput
@@ -124,51 +115,20 @@ getModelFit <- function(data,
   }
 }
 
-fitModel <- function(prepData,
-                     modelParameters,
-                     progressMessage = "Calculating Model",
-                     isCheck = FALSE){
-
-  K <- modelParameters$K
-  smoothConst <- modelParameters$smoothConst
-  burnin <- modelParameters$burnin
-  iter <- modelParameters$iter
-  nChains <- modelParameters$nChains
-  sdVar <- modelParameters$sdVar
-  const <- modelParameters$const
-
-  tryCatchWithMessage(
-    fitPlotRModelMC(
-      prepData,
-      K = K,
-      burnin = burnin,
-      iter = iter,
-      penalty = const,
-      smoothConst = smoothConst,
-      nChains = nChains,
-      sdVar = sdVar,
-      progressMessage = progressMessage,
-      isCheck = isCheck
-    )
-  )
-}
-
 fitPlotRModelMC <- function(data,
                             K = 24, burnin = 1000,
                             iter = 24000, penalty = 2,
                             smoothConst = 1,
                             nChains = 4,
                             sdVar = FALSE,
-                            progressMessage = "Calculating Model",
-                            isCheck = FALSE){
-
+                            progressMessage = "Calculating Model"){
   n <- nrow(data)
   if (n < K) {
-    shinyalert("Not enough rows for running the model.",
-               "Please use more data rows, increase the sd for outlier removal or
-                     decrease the number of basis functions.",
-               type = "error")
-    return(NULL)
+    stop("Not enough rows for running the model. Please use more data rows, increase the sd for outlier removal or decrease the number of basis functions.")
+  }
+
+  if (n < 8) {
+    warning("Estimation of the standard deviation / the SEM with nonlinear models may not be reliable for sample sizes below 8.")
   }
 
   ret <- lapply(1:nChains, function(x){
@@ -178,8 +138,7 @@ fitPlotRModelMC <- function(data,
                   smoothConst = smoothConst,
                   nChains = x,
                   sdVar = sdVar,
-                  progressMessage = progressMessage,
-                  isCheck = isCheck)
+                  progressMessage = progressMessage)
   })
   res <- ret[[1]]
   res$beta <- do.call("rbind", lapply(1:length(ret), function(x) ret[[x]]$beta))
@@ -195,8 +154,7 @@ fitPlotRModel <- function(data,
                           smoothConst = 1,
                           nChains = 4,
                           sdVar = FALSE,
-                          progressMessage = "Calculating Model",
-                          isCheck = FALSE){
+                          progressMessage = "Calculating Model"){
 
   data$Date4 <- data$standardizedX
 
@@ -371,7 +329,7 @@ fitPlotRModel <- function(data,
     return(betamc)
   }
 
-  if (isCheck) {
+  if (!isRunning()) {
     for ( k in 1:10) {
       j <- seq(1, iter, iter / 10)[k]
       MCMCplotR(start = j, iter = j + iter / 10 - 1)