+++ R/03-stylePlot.R
@@ -8,7 +8,8 @@ stylePlotUI <- function(id, title) {
            value = id,
            fluidRow(
              sidebarPanel(
-               style = "position:fixed; width:20%; max-width:350px; overflow-y:auto; height:88%",
+               style = "position:fixed; width:23%; max-width:500px; overflow-y:auto; height:88%",
+               width = 3,
                selectInput(
                  ns("activePlot"),
                  label = "Select a saved plot",
@@ -206,11 +207,28 @@ stylePlotUI <- function(id, title) {
                      min = 0,
                      max = 20,
                      value = 2
+                   ),
+                   conditionalPanel(
+                     condition = "input.lineName == 'modelUncertainty'",
+                     ns = ns,
+                   colourInput(
+                     inputId = ns("uncertaintyBandColor"),
+                     label = "Uncertainty Band Color",
+                     value = '#1D60BD'
+                   ),
+                   numericInput(
+                     inputId = ns("uncertaintyBandOpacity"),
+                     label = "Uncertainty Band Opacity",
+                     min = 0,
+                     max = 1,
+                     value = 0.2,
+                     step = 0.1
                    )
                  )
                )
              )
-           ))
+            )
+          ))
 }
 
 #' @export
@@ -297,7 +315,11 @@ stylePlot <- function(input, output, session, savedData) {
     updateColourInput(session, "lineColor", value = activePlotStyle[[getLastSelected(lastSelected$line, "predictionLine")]]$color)
     updateSelectInput(session, "lineType", selected = activePlotStyle[[getLastSelected(lastSelected$line, "predictionLine")]]$lineType)
     updateSliderInput(session, "lineWidth", value = activePlotStyle[[getLastSelected(lastSelected$line, "predictionLine")]]$lineWidth)
-    updateCheckboxInput(session, "hideHide", value = activePlotStyle[[getLastSelected(lastSelected$line, "predictionLine")]]$hideHide)
+    updateCheckboxInput(session, "hideHide", value = activePlotStyle[[getLastSelected(lastSelected$line, "predictionLine")]]$hide)
+
+    # uncertainty band
+    updateColourInput(session, "uncertaintyBandColor", value = activePlotStyle$modelUncertainty$bandColor)
+    updateNumericInput(session, "uncertaintyBandOpacity", value = activePlotStyle$modelUncertainty$bandOpacity)
   })
 
   # activePlotStyle(changed label, point, line) >> inputs ####
@@ -370,15 +392,23 @@ stylePlot <- function(input, output, session, savedData) {
         color = input$lineColor,
         lineType = as.numeric(input$lineType),
         lineWidth = input$lineWidth,
-        hide = input$lineHide
+        hide = input$lineHide,
+        bandColor = input$uncertaintyBandColor,
+        bandOpacity = input$uncertaintyBandOpacity
       )
   })
 
   # render plot ####
   values <- reactiveValues(plot = NULL)
   output$styledPlot <- renderPlot({
+    validate(
+      need(input$activePlot, "Select a plot ...")
+    )
     req(input$activePlot)
-    req(activePlotValues$predictedData$evenlyOnX)
+    validate(
+      need(!is.null(activePlotValues$defaultXRange), "Data not valid ...")
+    )
+
     makeSinglePlot(
       reactiveValuesToList(activePlotValues),
       reactiveValuesToList(activePlotStyle)
@@ -398,28 +428,28 @@ stylePlot <- function(input, output, session, savedData) {
   # delete plot ####
   callModule(deletePlot, "deletingPlot", savedData = savedData)
 
-  dataFun <- reactive({
-    req(input$activePlot)
-    req(activePlotValues$modelData)
-    function(xVar, quantile) {
-      savedData <-
-        predictPipe(
-          plotRModel = activePlotValues$modelData$modelOutput,
-          xCol = activePlotValues$prepData$X,
-          xVar = xVar,
-          yName = activePlotValues$ySelection$colNames$colName1,
-          quantile = quantile
-        )
-      return(savedData)
-    }
-  })
+  # dataFun <- reactive({
+  #   req(input$activePlot)
+  #   req(activePlotValues$modelData)
+  #   function(xVar, quantile) {
+  #     savedData <-
+  #       predictPipe(
+  #         plotRModel = activePlotValues$modelData$modelOutput,
+  #         xCol = activePlotValues$prepData$X,
+  #         xVar = xVar,
+  #         yName = activePlotValues$ySelection$colNames$colName1,
+  #         quantile = quantile
+  #       )
+  #     return(savedData)
+  #   }
+  # })
 
   #callModule(dataExport, "exportData", dat = dataFun, filename = "modelData")
   callModule(plotExport, "export", reactive(values$plot))
 }
 
 getLastSelected <- function(lastSelected, default) {
-  if_else(is.null(lastSelected),
+  ifelse(is.null(lastSelected),
           default,
           lastSelected)
 }